Move CommentGenerator classes into plugin
